requirements.txt
asgiref==3.6.0
certifi==2022.12.7
distlib==0.3.6
Django==4.1.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.9.0
pipenv==2022.12.19
platformdirs==2.6.2
psycopg2==2.9.5
PyJWT==2.6.0
python-decouple==3.7
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
virtualenv==20.17.1
virtualenv-clone==0.5.7
